refactor: move templates from sv to cmd/git-sv
issue: #40
