+++ sv/formatter_test.go
@@ -2,12 +2,15 @@ package sv
 
 import (
 	"bytes"
+	"os"
 	"testing"
 	"text/template"
 	"time"
 
 	"github.com/Masterminds/semver/v3"
 )
+
+var templatesFS = os.DirFS("../cmd/git-sv/resources/templates")
 
 var dateChangelog = `## v1.0.0 (2020-05-01)
 `
@@ -57,7 +60,7 @@ func TestOutputFormatterImpl_FormatReleaseNote(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got, err := NewOutputFormatter().FormatReleaseNote(tt.input)
+			got, err := NewOutputFormatter(templatesFS).FormatReleaseNote(tt.input)
 			if got != tt.want {
 				t.Errorf("OutputFormatterImpl.FormatReleaseNote() = %v, want %v", got, tt.want)
 			}
@@ -88,6 +91,30 @@ func fullReleaseNote(tag string, date time.Time) ReleaseNote {
 	return releaseNote(v, tag, date, sections, []string{"break change message"})
 }
 
+func Test_checkTemplatesExecution(t *testing.T) {
+	tpls := template.Must(template.New("templates").ParseFS(templatesFS, "*"))
+	tests := []struct {
+		template  string
+		variables interface{}
+	}{
+		{"changelog-md.tpl", changelogVariables("v1.0.0", "v1.0.1")},
+		{"releasenotes-md.tpl", releaseNotesVariables("v1.0.0")},
+	}
+	for _, tt := range tests {
+		t.Run(tt.template, func(t *testing.T) {
+			var b bytes.Buffer
+			err := tpls.ExecuteTemplate(&b, tt.template, tt.variables)
+			if err != nil {
+				t.Errorf("invalid template err = %v", err)
+				return
+			}
+			if len(b.Bytes()) <= 0 {
+				t.Errorf("empty template")
+			}
+		})
+	}
+}
+
 func releaseNotesVariables(release string) releaseNoteTemplateVariables {
 	return releaseNoteTemplateVariables{
 		Release: release,
@@ -109,47 +136,4 @@ func changelogVariables(releases ...string) []releaseNoteTemplateVariables {
 	}
 	return variables
 
-}
-
-func Test_checkTemplatesFiles(t *testing.T) {
-	tests := []string{
-		"resources/templates/changelog-md.tpl",
-		"resources/templates/releasenotes-md.tpl",
-	}
-	for _, tt := range tests {
-		t.Run(tt, func(t *testing.T) {
-			got, err := defaultTemplatesFS.ReadFile(tt)
-			if err != nil {
-				t.Errorf("missing template error = %v", err)
-				return
-			}
-			if len(got) <= 0 {
-				t.Errorf("empty template")
-			}
-		})
-	}
-}
-
-func Test_checkTemplatesExecution(t *testing.T) {
-	tpls := template.Must(template.New("templates").ParseFS(defaultTemplatesFS, "resources/templates/*"))
-	tests := []struct {
-		template  string
-		variables interface{}
-	}{
-		{"changelog-md.tpl", changelogVariables("v1.0.0", "v1.0.1")},
-		{"releasenotes-md.tpl", releaseNotesVariables("v1.0.0")},
-	}
-	for _, tt := range tests {
-		t.Run(tt.template, func(t *testing.T) {
-			var b bytes.Buffer
-			err := tpls.ExecuteTemplate(&b, tt.template, tt.variables)
-			if err != nil {
-				t.Errorf("invalid template err = %v", err)
-				return
-			}
-			if len(b.Bytes()) <= 0 {
-				t.Errorf("empty template")
-			}
-		})
-	}
 }
